Move the degree computation into a background thread
